% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_count.R
\name{add_count}
\alias{add_count}
\title{for each input get the count of dependent nodes.}
\usage{
add_count(plan_dt, node)
}
\arguments{
\item{plan_dt}{`data.table` object describing a simplan.}

\item{node}{Name of the node.}
}
\description{
for each input get the count of dependent nodes.
}
