% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_iters.R
\name{add_iters}
\alias{add_iters}
\title{Add iteration indexing for for loops.}
\usage{
add_iters(string, inputs)
}
\arguments{
\item{string}{string of parameters and values}

\item{inputs}{vector of names of input nodes.}
}
\description{
Add iteration indexing for for loops.
}
