% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listify.R
\name{combine}
\alias{combine}
\title{Internal function for converting an R-list into an R-object of the specified type.}
\usage{
combine(list, type)
}
\arguments{
\item{list}{R List object created by `listify()`.}

\item{type}{Type of simplan array (currently only matrix and vector accepted)).}
}
\description{
Internal function for converting an R-list into an R-object of the specified type.
}
