% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_r_target.Internal.R
\name{construct_r_target.Internal}
\alias{construct_r_target.Internal}
\title{Construct R targets code for an expression.}
\usage{
construct_r_target.Internal(plan, node, add_nsims = FALSE)
}
\arguments{
\item{plan}{Simplan plan.}

\item{node}{Name of the node.}

\item{add_nsims}{Add the n-sims directly.}
}
\description{
Construct R targets code for an expression.
}
