% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listify.R
\name{express_array_param}
\alias{express_array_param}
\title{Internal function for converting a simplan array string into a R-expression (still string.)}
\usage{
express_array_param(string)
}
\arguments{
\item{string}{string}
}
\description{
Internal function for converting a simplan array string into a R-expression (still string.)
}
