% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_variables.Internal.R
\name{extract_variables.Internal}
\alias{extract_variables.Internal}
\title{Extract the variables from an expression.}
\usage{
extract_variables.Internal(x)
}
\arguments{
\item{x}{A character string representing an expression.}
}
\description{
Extract the variables from an expression.
}
