% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_variables.Internal.R
\name{extract_variables_2.Internal}
\alias{extract_variables_2.Internal}
\title{Extract the variables from multiple expressions and convert to a character vector.}
\usage{
extract_variables_2.Internal(x)
}
\arguments{
\item{x}{List of character strings representing expressions.}
}
\description{
Extract the variables from multiple expressions and convert to a character vector.
}
