% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_arrays.R
\name{get_first_bracket_end}
\alias{get_first_bracket_end}
\title{Identify the index of the closing bracket}
\usage{
get_first_bracket_end(x)
}
\arguments{
\item{x}{Character string with brackets.}
}
\description{
Identify the index of the closing bracket
}
