% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node_rank.R
\name{get_node_count.Internal}
\alias{get_node_count.Internal}
\title{Recursively determine the number of inputs (direct and indirect) into a node.}
\usage{
get_node_count.Internal(
  node,
  node_inputs,
  nodes = names(node_inputs),
  remove_selfreference = TRUE
)
}
\arguments{
\item{node}{Node in question}

\item{node_inputs}{A list of character vectors with each of the inputs into the node.}

\item{nodes}{Names of the nodes in the list.}

\item{remove_selfreference}{Remove self-references?}
}
\description{
Recursively determine the number of inputs (direct and indirect) into a node.
}
