% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node_rank.R
\name{get_node_rank}
\alias{get_node_rank}
\title{Determine the build order for a set of nodes}
\usage{
get_node_rank(node_inputs, nodes = names(node_inputs))
}
\arguments{
\item{node_inputs}{A list of character vectors with each of the inputs into the node.}

\item{nodes}{Names of the nodes in the list.}
}
\description{
Determine the build order for a set of nodes
}
\examples{
node_inputs <- list(
  alpha = c("pi","f", "alpha"),
  beta = c("alpha","epsilon"),
  delta = c("d","beta"),
  epsilon = c("2")
)
simplan:::get_node_rank(node_inputs)
}
