% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listify.R
\name{listify}
\alias{listify}
\title{Internal function for expressing heirarchical string lists as R list() objects.}
\usage{
listify(string)
}
\arguments{
\item{string}{string}
}
\description{
Internal function for expressing heirarchical string lists as R list() objects.
}
