% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_as_rtargets.R
\name{output_as_rtargets}
\alias{output_as_rtargets}
\title{Output a simplan into an R script.}
\usage{
output_as_rtargets(file = "_targets.R", plan, add_nsims = TRUE)
}
\arguments{
\item{file}{Output file path.}

\item{plan}{Simplan plan.}

\item{add_nsims}{Add the n-sims directly.}
}
\description{
Output a simplan into an R script.
}
