% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_simulation_DAG}
\alias{plot_simulation_DAG}
\title{Plot the directed acyclic graph of the nodes.}
\usage{
plot_simulation_DAG(plan)
}
\arguments{
\item{plan}{A baked simplan}
}
\description{
Plot the directed acyclic graph of the nodes.
}
