% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_distributions.R
\name{rbootstrap}
\alias{rbootstrap}
\title{Sample a bootstrap statistic from a finite set.}
\usage{
rbootstrap(n, data, statistic)
}
\arguments{
\item{n}{Number of "draws".}

\item{data}{Data vector to draw resamples.}

\item{statistic}{Statistic to run.}
}
\description{
Sample a bootstrap statistic from a finite set.
}
