% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_distributions.R
\name{rconstant}
\alias{rconstant}
\title{Constant distribution (dirac).}
\usage{
rconstant(n, location)
}
\arguments{
\item{n}{Number of "draws".}

\item{location}{Location of spike/only value with non-zero probability.}
}
\description{
Constant distribution (dirac).
}
