% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_distributions.R
\name{rdiscrete}
\alias{rdiscrete}
\title{Sample from a finite set.}
\usage{
rdiscrete(n, locations, frequencies = NULL)
}
\arguments{
\item{n}{Number of "draws".}

\item{locations}{Location of spikes/only values with non-zero probability.}

\item{frequencies}{Frequencies/probabilities of the spikes.}
}
\description{
Sample from a finite set.
}
