% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_distributions.R
\name{rmvnorm}
\alias{rmvnorm}
\title{Multivariate normal:}
\usage{
rmvnorm(n, mean, covariance)
}
\arguments{
\item{n}{Number of draws.}

\item{mean}{vector mean.}

\item{covariance}{matrix for covariance.}
}
\description{
Multivariate normal:
}
