% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_distributions.R
\name{runiform}
\alias{runiform}
\title{Uniform distribution}
\usage{
runiform(n, lower, upper)
}
\arguments{
\item{n}{Number of draws.}

\item{lower}{Lower limit.}

\item{upper}{Upper limit.}
}
\description{
Uniform distribution
}
