% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_arrays.R
\name{translate_arrays}
\alias{translate_arrays}
\title{Identify all substrings of the form MATRIX|VECTOR(...)}
\usage{
translate_arrays(string)
}
\arguments{
\item{string}{String of simplan code to be translated into R.}
}
\description{
Identify all substrings of the form MATRIX|VECTOR(...)
}
